third_party/blink/web_tests/svg/dynamic-updates/SVGMarkerElement-dom-refY-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'refY' attribute of the SVGMarkerElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS markerElement.getAttribute('refY') is "500"
PASS markerElement.getAttribute('refY') is "5"
PASS successfullyParsed is true

TEST COMPLETE
